internal/olm: move GetInstalledVersion() to internal/olm/client
